+++ nextgisweb/webmap/amd/ngw-webmap/Display.js
@@ -29,18 +29,21 @@ define([
     "./ui/react-panel",
     "./ui/react-webmap-tabs",
     // tools
+    "@nextgisweb/webmap/map-controls",
     "./tool/Zoom",
     "./tool/Measure",
     "./tool/Identify",
     "./tool/ViewerInfo",
     "./tool/Swipe",
+    // Tiny display
+    "ngw-webmap/controls/LinkToMainMap",
     // panels
     "@nextgisweb/webmap/panel/layers",
     // utils
     "./utils/URL",
     // settings
     "@nextgisweb/pyramid/i18n!",
-    "@nextgisweb/pyramid/settings!",
+    "@nextgisweb/pyramid/settings!webmap",
     "dojo/text!./template/Display.hbs",
     // template
     "dijit/layout/BorderContainer",
@@ -77,11 +80,13 @@ define([
     MapStatesObserver,
     reactPanel,
     ReactWebMapTabs,
+    MapControls,
     ToolZoom,
     ToolMeasure,
     Identify,
     ToolViewerInfo,
     ToolSwipe,
+    LinkToMainMap,
     LayersPanelModule,
     URL,
     { gettext, renderTemplate },
@@ -174,6 +179,7 @@ define([
         emptyModeURLValue: "none",
 
         webmapStore: undefined,
+        tinyConfig: undefined,
 
         constructor: function (options) {
             declare.safeMixin(this, options);
@@ -199,20 +205,7 @@ define([
 
             this.clientSettings = settings;
 
-            const activePanelKey = this._urlParams[this.modeURLParam];
-            const onChangePanel = (panel) => {
-                if (panel) {
-                    URL.setURLParam(this.modeURLParam, panel.name);
-                } else {
-                    URL.setURLParam(this.modeURLParam, this.emptyModeURLValue);
-                }
-            };
-            this.panelsManager = new PanelsManager.default(
-                this,
-                activePanelKey,
-                onChangePanel
-            );
-            this._buildPanels();
+            this._buildPanelsManager();
 
             this.tabContainer = new ReactWebMapTabs({ display: this });
 
@@ -360,6 +353,7 @@ define([
                 navigation: this.navigationMenuPane.domNode,
             };
             this.panelsManager.initDomElements(domElements);
+            this._handleTinyDisplayMode();
 
             this._postCreateDeferred.resolve();
         },
@@ -470,54 +464,7 @@ define([
                 }),
             });
 
-            this._mapAddControls([
-                new ol.control.Zoom({
-                    zoomInLabel: domConstruct.create("span", {
-                        class: "ol-control__icon",
-                        innerHTML: icon.html({ glyph: "add" }),
-                    }),
-                    zoomOutLabel: domConstruct.create("span", {
-                        class: "ol-control__icon",
-                        innerHTML: icon.html({ glyph: "remove" }),
-                    }),
-                    zoomInTipLabel: gettext("Zoom in"),
-                    zoomOutTipLabel: gettext("Zoom out"),
-                    target: widget.leftTopControlPane,
-                }),
-                new ol.control.Attribution({
-                    tipLabel: gettext("Attributions"),
-                    target: widget.rightBottomControlPane,
-                    collapsible: false,
-                }),
-                new ol.control.ScaleLine({
-                    target: widget.rightBottomControlPane,
-                    units: settings.units,
-                    minWidth: 48,
-                }),
-                new InfoScale({
-                    display: widget,
-                    target: widget.rightBottomControlPane,
-                }),
-                new InitialExtent({
-                    display: widget,
-                    target: widget.leftTopControlPane,
-                    tipLabel: gettext("Initial extent"),
-                }),
-                new MyLocation({
-                    display: widget,
-                    target: widget.leftTopControlPane,
-                    tipLabel: gettext("Locate me"),
-                }),
-                new ol.control.Rotate({
-                    tipLabel: gettext("Reset rotation"),
-                    target: widget.leftTopControlPane,
-                    label: domConstruct.create("span", {
-                        class: "ol-control__icon",
-                        innerHTML: icon.html({ glyph: "arrow_upward" }),
-                    }),
-                }),
-                widget.mapToolbar,
-            ]);
+            MapControls.buildControls(this);
 
             // Resize OpenLayers Map on container resize
             aspect.after(this.mapPane, "resize", function () {
@@ -553,9 +500,9 @@ define([
                     } catch (err) {
                         console.warn(
                             "Can't initialize layer [" +
-                                baseOptions.keyname +
-                                "]: " +
-                                err
+                            baseOptions.keyname +
+                            "]: " +
+                            err
                         );
                     }
 
@@ -639,10 +586,10 @@ define([
                 queryOptions: { deep: true },
                 sort: widget.config.drawOrderEnabled
                     ? [
-                          {
-                              attribute: "position",
-                          },
-                      ]
+                        {
+                            attribute: "position",
+                        },
+                    ]
                     : null,
                 onItem: function (item) {
                     widget._onNewStoreItem(item, visibleStyles);
@@ -697,34 +644,6 @@ define([
         },
 
         _toolsSetup: function () {
-            this.mapToolbar.items.addTool(
-                new ToolZoom({ display: this, out: false }),
-                "zoomingIn"
-            );
-            this.mapToolbar.items.addTool(
-                new ToolZoom({ display: this, out: true }),
-                "zoomingOut"
-            );
-
-            this.mapToolbar.items.addTool(
-                new ToolMeasure({ display: this, type: "LineString" }),
-                "measuringLength"
-            );
-            this.mapToolbar.items.addTool(
-                new ToolMeasure({ display: this, type: "Polygon" }),
-                "measuringArea"
-            );
-
-            this.mapToolbar.items.addTool(
-                new ToolSwipe({ display: this, orientation: "vertical" }),
-                "swipeVertical"
-            );
-
-            this.mapToolbar.items.addTool(
-                new ToolViewerInfo({ display: this }),
-                "~viewerInfo"
-            );
-
             this.identify = new Identify({ display: this });
 
             this.mapStates.addState("identifying", this.identify);
@@ -941,6 +860,156 @@ define([
                 });
         },
 
+        _handleTinyDisplayMode: function () {
+            if (!this.isTinyMode()) {
+                return;
+            }
+
+            this.domNode.classList.add("tiny");
+
+            all([
+                this._layersDeferred,
+                this._mapDeferred,
+                this._postCreateDeferred,
+                this.panelsManager.panelsReady.promise,
+            ])
+                .then(() => {
+                    this._buildTinyPanels();
+                    this._addLinkToMainMap();
+                    this._handlePostMessage();
+                })
+                .then(undefined, function (err) {
+                    console.error(err);
+                });
+        },
+
+        _buildTinyPanels: function () {
+            if (!this.panelsManager.getPanelsCount()) {
+                return;
+            }
+
+            this.domNode.classList.add("tiny-panels");
+            const activePanel = this.panelsManager.getActivePanelName();
+            if (!activePanel) {
+                return;
+            }
+            this.panelsManager.deactivatePanel();
+            this.panelsManager.activatePanel(activePanel);
+        },
+
+        _addLinkToMainMap: function () {
+            if (this._urlParams.linkMainMap !== "true") {
+                return;
+            }
+            this.map.olMap.addControl(
+                new LinkToMainMap({
+                    url: this.tinyConfig.mainDisplayUrl,
+                    target: this.rightTopControlPane,
+                    tipLabel: gettext("Open full map"),
+                })
+            );
+        },
+
+        /**
+         * Generate window `message` events to listen from iframe
+         * @example
+         * window.addEventListener('message', function(evt) {
+         *    var data = evt.data;
+         *    if (data.event === 'ngMapExtentChanged') {
+         *        if (data.detail === 'zoom') {
+         *        } else if (data.detail === 'move') {
+         *        }
+         *        // OR
+         *        if (data.detail === 'position') {}
+         *    }
+         * }, false);
+         */
+        _handlePostMessage: function () {
+            var widget = this;
+            var parent = window.parent;
+            if (
+                this._urlParams.events === "true" &&
+                parent &&
+                parent.postMessage
+            ) {
+                var commonOptions = {
+                    event: "ngMapExtentChanged",
+                };
+                var parsePosition = function (pos) {
+                    return {
+                        zoom: pos.zoom,
+                        lat: pos.center[1],
+                        lon: pos.center[0],
+                    };
+                };
+                widget.map.watch(
+                    "position",
+                    function (name, oldPosition, newPosition) {
+                        oldPosition = oldPosition
+                            ? parsePosition(oldPosition)
+                            : {};
+                        newPosition = parsePosition(newPosition);
+                        // set array of position part to compare between old and new state
+                        var events = [
+                            { params: ["lat", "lon"], name: "move" },
+                            { params: ["zoom"], name: "zoom" },
+                        ];
+                        var transformPosition = widget.map.getPosition(
+                            widget.lonlatProjection
+                        );
+                        // prepare to send transform position
+                        commonOptions.data = parsePosition(transformPosition);
+                        array.forEach(events, function (event) {
+                            var isChange = array.some(
+                                event.params,
+                                function (p) {
+                                    return oldPosition[p] !== newPosition[p];
+                                }
+                            );
+                            if (isChange) {
+                                commonOptions.detail = event.name;
+                                // message should be a string to work correctly with all browsers and systems
+                                parent.postMessage(
+                                    JSON.stringify(commonOptions),
+                                    "*"
+                                );
+                            }
+                        });
+                        // on any position change
+                        commonOptions.detail = name;
+                        parent.postMessage(JSON.stringify(commonOptions), "*");
+                    }
+                );
+            }
+        },
+
+        _buildPanelsManager: function () {
+            const activePanelKey = this._urlParams[this.modeURLParam];
+            const onChangePanel = (panel) => {
+                if (panel) {
+                    URL.setURLParam(this.modeURLParam, panel.name);
+                } else {
+                    URL.setURLParam(this.modeURLParam, this.emptyModeURLValue);
+                }
+            };
+
+            let allowPanels;
+            if (this.isTinyMode()) {
+                allowPanels = this._urlParams.panels
+                    ? this._urlParams.panels.split(",")
+                    : [];
+            }
+
+            this.panelsManager = new PanelsManager.default(
+                this,
+                activePanelKey,
+                allowPanels,
+                onChangePanel
+            );
+
+            this._buildPanels();
+        },
+
         _buildPanels: function () {
             const promises = all([
                 this._layersDeferred,
@@ -966,6 +1035,7 @@ define([
                     name: "layers",
                     order: 10,
                     menuIcon: "material-layers",
+                    applyToTinyMap: true,
                 },
             });
 
@@ -976,6 +1046,7 @@ define([
                     name: "search",
                     order: 20,
                     menuIcon: "material-search",
+                    applyToTinyMap: true,
                 },
             });
 
@@ -1000,6 +1071,7 @@ define([
                         name: "bookmark",
                         order: 50,
                         menuIcon: "material-bookmark",
+                        applyToTinyMap: true,
                     },
                 };
                 resolve(panel);
@@ -1017,6 +1089,7 @@ define([
                         name: "info",
                         order: 40,
                         menuIcon: "material-info-outline",
+                        applyToTinyMap: true,
                     },
                 };
                 resolve(panel);
@@ -1087,6 +1160,7 @@ define([
                     name: "annotation",
                     order: 30,
                     menuIcon: "material-message",
+                    applyToTinyMap: true,
                 },
             };
             resolve(panel);
@@ -1097,6 +1171,14 @@ define([
             topic.publish("feature.highlight", {
                 olGeometry: geometry,
             });
+        },
+
+        getUrlParams: function () {
+            return this._urlParams;
+        },
+
+        isTinyMode: function () {
+            return this.tinyConfig !== undefined;
         },
     });
 });