+++ nextgisweb/webmap/nodepkg/map-controls/index.ts
@@ -1,0 +1,4 @@
+/** @entrypoint */
+import { buildControls, getControls } from "./map-controls";
+
+export { buildControls, getControls };