+++ nextgisweb/webmap/nodepkg/type/DisplayConfig.ts
@@ -42,4 +42,5 @@ export interface DisplayConfig {
     webmapLegendVisible: string;
     drawOrderEnabled?: any;
     annotations: Annotations;
+    units: string;
 }