+++ nextgisweb/webmap/nodepkg/type/DojoDisplay.ts
@@ -1,5 +1,6 @@
 /// <reference types="dojo/dijit" />
 
+import type { Control as OlControl } from "ol/control";
 import type Feature from "ol/Feature";
 
 import type WebmapStore from "../store";
@@ -10,6 +11,7 @@ import type { MapStatesObserver } from "./MapState";
 import type { WebmapItem } from "./WebmapItem";
 import type { WebmapLayer } from "./WebmapLayer";
 import type { WebmapPlugin } from "./WebmapPlugin";
+import { DojoItem } from ".";
 
 interface DojoDisplayIdentifyPopup {
     widget?: DojoDisplayIdentify;
@@ -37,17 +39,47 @@ export interface FeatureHighlighter {
     unhighlightFeature: (filter: (feature: Feature) => boolean) => void;
 }
 
+export interface MapTool extends dijit._WidgetBase {
+    activate: () => void;
+    deactivate: () => void;
+}
+
+export interface MapToolbarItems {
+    addTool: (tool: MapTool, state: string, place?: HTMLElement) => void;
+    addSeparator: () => void;
+    addButton: (button: DojoItem, options: any) => void;
+}
+
+export interface MapToolbar extends OlControl {
+    items: MapToolbarItems;
+}
+
+export type MapControl = OlControl | dijit._WidgetBase;
+
 export interface DojoDisplay extends dijit._WidgetBase {
+    config: DisplayConfig;
     identify: DojoDisplayIdentify;
     featureHighlighter: FeatureHighlighter;
-    map: DisplayMap;
-    mapContainer: dijit.layout.BorderContainer;
-    displayProjection: string;
-    config: DisplayConfig;
+    getUrlParams: () => Record<string, string>;
+    isTinyMode: () => boolean;
+
     itemStore: CustomItemFileWriteStore;
     webmapStore: WebmapStore;
     mapStates: MapStatesObserver;
+
+    map: DisplayMap;
+    mapContainer: dijit.layout.BorderContainer;
+    displayProjection: string;
+
+    mapToolbar: MapToolbar;
+    _mapAddControls: (controls: MapControl[]) => void;
     _plugins: WebmapPlugin[];
+
+    leftTopControlPane: HTMLDivElement;
+    leftBottomControlPane: HTMLDivElement;
+    rightTopControlPane: HTMLDivElement;
+    rightBottomControlPane: HTMLDivElement;
+
     /**
      * @deprecated use webmapStore.getlayers() instead
      */